# Benchmark Reader and inline its requests.