--- a/benchmark/Bench.hs
+++ b/benchmark/Bench.hs
@@ -11,11 +11,13 @@ import Control.Monad (replicateM_)
 import Data.Monoid (Sum(..))
 import Gauge
 
+import qualified Bench.Reader as Reader
 import qualified Bench.NonDet as NonDet
 
 main :: IO ()
 main = defaultMain
-  [ NonDet.benchmark
+  [ Reader.benchmark
+  , NonDet.benchmark
   , bgroup "WriterC"
     [ bench "100"   $ whnf (run . execWriter @(Sum Int) . tellLoop) 100
     , bench "1000"  $ whnf (run . execWriter @(Sum Int) . tellLoop) 1000