--- a/fused-effects.cabal
+++ b/fused-effects.cabal
@@ -177,6 +177,7 @@ benchmark benchmark
   hs-source-dirs: benchmark
   main-is:        Bench.hs
   other-modules:
+    Bench.Reader
     Bench.NonDet
     Bench.NonDet.NQueens
   build-depends: