--- a/src/Control/Effect/Reader.hs
+++ b/src/Control/Effect/Reader.hs
@@ -38,6 +38,7 @@ import Control.Effect.Reader.Internal (Reader(..))
 -- @since 0.1.0.0
 ask :: Has (Reader r) sig m => m r
 ask = send (Ask pure)
+{-# INLINE ask #-}
 
 -- | Project a function out of the current environment value.
 --
@@ -48,6 +49,7 @@ ask = send (Ask pure)
 -- @since 0.1.0.0
 asks :: Has (Reader r) sig m => (r -> a) -> m a
 asks f = send (Ask (pure . f))
+{-# INLINE asks #-}
 
 -- | Run a computation with an environment value locally modified by the passed function.
 --
@@ -58,3 +60,4 @@ asks f = send (Ask (pure . f))
 -- @since 0.1.0.0
 local :: Has (Reader r) sig m => (r -> r) -> m a -> m a
 local f m = send (Local f m pure)
+{-# INLINE local #-}