--- a/src/Control/Effect/Reader/Labelled.hs
+++ b/src/Control/Effect/Reader/Labelled.hs
@@ -32,6 +32,7 @@ import           Control.Effect.Reader.Internal
 -- @since 1.0.2.0
 ask :: forall label r m sig . HasLabelled label (Reader r) sig m => m r
 ask = runUnderLabel @label R.ask
+{-# INLINE ask #-}
 
 -- | Project a function out of the current environment value.
 --
@@ -42,6 +43,7 @@ ask = runUnderLabel @label R.ask
 -- @since 1.0.2.0
 asks :: forall label r m a sig . HasLabelled label (Reader r) sig m => (r -> a) -> m a
 asks f = runUnderLabel @label (R.asks f)
+{-# INLINE asks #-}
 
 -- | Run a computation with an environment value locally modified by the passed function.
 --
@@ -52,3 +54,4 @@ asks f = runUnderLabel @label (R.asks f)
 -- @since 1.0.2.0
 local :: forall label r m a sig . HasLabelled label (Reader r) sig m => (r -> r) -> m a -> m a
 local f m = runUnderLabel @label (R.local f (UnderLabel m))
+{-# INLINE local #-}