@patrickt 2020-01-20
@ocharles Yeah, @robrix and I were looking at this. It’s unclear what’s going on, but I’m hesitant to dig in until we have concrete evidence that this is affecting real-world apps.