@robrix 2020-01-21
@patrickt: I’d rather not close out #345 as fixed until we’ve `INLINE`d _every_ effect’s operators. I don’t think we need to benchmark all of them, tho; would you mind `INLINE`ing the others?