@robrix 2020-01-21
Also, I don’t think the `INLINE` vs. `INLINABLE` thing is really conclusive; given the variance, I think it could go either way. Unlike the `-O2` thing, I think this _does_ warrant some investigation of the core; what’s your take?