@robrix 2020-01-21
(It occurs to me that we could benchmark them against each other directly by duplicating the constructors in separate modules or something, too; that’d be another tactic we could try.)