@robrix 2020-03-03
Upon reflection, I’m pretty ok with `INLINE`. These are tiny definitions and I think it’s reasonable to be opinionated about inlining them.